+++ backend/models/withdrawalrequest.js
@@ -51,9 +51,7 @@ module.exports = (sequelize, DataTypes) => {
         get() {
           const amount = this.getDataValue('amount');
           const currency = this.getDataValue('currency');
-          if (
-            (['usd', 'USD', '$'].includes(currency)) && amount
-          ) {
+          if (['usd', 'USD', '$'].includes(currency) && amount) {
             return '$' + amount;
           } else {
             return amount;
@@ -317,6 +315,193 @@ module.exports = (sequelize, DataTypes) => {
       subQuery: false,
     });
     return result.count;
+  };
+  //check approved withdrawal request
+  WithdrawalRequest.withdrawalRequestAmountValidation = async (
+    payment_method_details,
+    withdrawal_amount,
+    member_id,
+    member
+  ) => {
+    const { MemberTransaction } = require('../models/index');
+    //check all conditions for amount
+    if (
+      parseFloat(payment_method_details.minimum_amount) > 0 &&
+      withdrawal_amount < parseFloat(payment_method_details.minimum_amount)
+    ) {
+      return {
+        member_status: false,
+        member_message:
+          'Amount must be more than $' + payment_method_details.min_amount,
+      };
+    }
+    if (
+      parseFloat(payment_method_details.maximum_amount) > 0 &&
+      withdrawal_amount > parseFloat(payment_method_details.maximum_amount)
+    ) {
+      return {
+        member_status: false,
+        member_message:
+          'Amount must be less than $' + payment_method_details.max_amount,
+      };
+    }
+    if (
+      parseFloat(payment_method_details.fixed_amount) > 0 &&
+      withdrawal_amount !== parseFloat(payment_method_details.fixed_amount)
+    ) {
+      return {
+        member_status: false,
+        member_message:
+          'Amount must be fixed to $' + payment_method_details.max_amount,
+      };
+    }
+
+    if (withdrawal_amount > parseFloat(member.member_amounts[0].amount)) {
+      return {
+        member_status: false,
+        member_message: 'Please check your balance',
+      };
+    }
+
+    let total_approved_amount = 0;
+    let total_pending_amount = 0;
+
+    //Start - check pending withdrawal request
+    let pending_withdrawal_req_amount = await WithdrawalRequest.findOne({
+      // logging: console.log,
+      attributes: [
+        [
+          sequelize.fn('SUM', sequelize.col('WithdrawalRequest.amount')),
+          'total',
+        ],
+      ],
+      where: {
+        status: 'pending',
+        member_id: member_id,
+      },
+      include: {
+        model: MemberTransaction,
+        attributes: ['id'],
+        where: {
+          status: { [Op.ne]: 2 },
+        },
+        required: false,
+      },
+    });
+    total_pending_amount = pending_withdrawal_req_amount.dataValues.total
+      ? parseFloat(pending_withdrawal_req_amount.dataValues.total)
+      : 0;
+    if (
+      member.member_amounts[0].amount <
+      total_pending_amount + withdrawal_amount
+    ) {
+      return {
+        member_status: false,
+        member_message:
+          'You already have pending withdrawal requests. This request might exceed your balance. Please contact to admin.',
+      };
+    }
+    //End - check pending withdrawal request
+
+    //Start - check approved withdrawal request
+    let approved_withdrawal_req_amount = await WithdrawalRequest.findOne({
+      // logging: console.log,
+      attributes: [
+        [
+          sequelize.fn('SUM', sequelize.col('WithdrawalRequest.amount')),
+          'total',
+        ],
+      ],
+      where: {
+        status: 'approved',
+        member_id: member_id,
+      },
+      include: {
+        model: MemberTransaction,
+        attributes: ['id'],
+        where: {
+          status: { [Op.ne]: 2 },
+        },
+        required: true,
+      },
+    });
+    total_approved_amount = approved_withdrawal_req_amount.dataValues.total
+      ? parseFloat(approved_withdrawal_req_amount.dataValues.total)
+      : 0;
+    if (
+      member.member_amounts[0].amount <
+      total_pending_amount + total_approved_amount + withdrawal_amount
+    ) {
+      return {
+        member_status: false,
+        member_message:
+          'You already have some approved withdrawal requests which are still under process. This request might exceed your balance. Please contact to admin.',
+      };
+    }
+    //End - check approved withdrawal request
+  };
+  //Withdrawal form validation
+  WithdrawalRequest.withdrawalRequestFieldValidation = async (
+    payment_method_details
+  ) => {
+    //Start - Withdrawal form validation
+    var payment_field = [];
+    var member_payment_info = [];
+    for (const option of payment_method_details.PaymentMethodFieldOptions) {
+      var field_name = option.field_name.toLowerCase().replaceAll(' ', '');
+      const option_arr = ['email', 'phone', 'username'];
+      if (option_arr.includes(field_name)) {
+        payment_field.push(request_data[field_name]);
+        member_payment_info.push({
+          field_name: field_name,
+          field_value: request_data[field_name],
+        });
+      }
+      if (request_data[field_name] === '') {
+        return {
+          member_status: false,
+          member_message: 'Please enter ' + option.field_name.toLowerCase(),
+        };
+      }
+      if (option.field_type !== 'input') {
+        var regex = '';
+        if (option.field_type === 'email')
+          regex = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
+        if (option.field_type === 'number')
+          regex = /^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/;
+
+        if (!regex.test(request_data[field_name])) {
+          // if (!request_data.payment_field.match(regex)) {
+          return {
+            member_status: false,
+            member_message:
+              'Please enter valid ' + option.field_name.toLowerCase(),
+          };
+        }
+      }
+    }
+    //End - Withdrawal form validation
+
+    return { member_status: true, payment_field, member_payment_info };
+  };
+  //check for same payment info
+  WithdrawalRequest.checkIfSamePaymentInfo = async (
+    payment_field,
+    member_id
+  ) => {
+    let check_same_acc = await WithdrawalRequest.count({
+      where: {
+        payment_email: payment_field,
+        member_id: { [Op.ne]: member_id },
+      },
+    });
+    if (check_same_acc > 0) {
+      return {
+        member_status: false,
+        member_message:
+          'This payment info has already been used by another account, please reach out to our admin',
+      };
+    }
   };
   return WithdrawalRequest;
 };
